Move Decodable extension to its own file
